stop importing secondary (slashy) systems
I'm thinking that slashy systems are easier to implement using
overrideLispAttrs->systems.

Some notes:

1. These systems are really part of the primary system. To me this translates
naturally to that they are all part of one Nix package.

2. Quicklisp doesn't include subsystems in dependency lists of other systems. It
can do that, because the system definition file for the parent system is enough
to build its subsystems as needed.

3. Subsystems might have additional dependencies to the primary system. Though
there is the possibility to merge the dependencies of all subsystems into the
primary system, this may encounter circular dependencies sometimes.

4. There is a combinatoric explosion of combinations of enabled subsystems. I
don't know currently how to test each combination, keeping in mind the above, in
a sane way

5. With overriding the systems parameter to build-asdf-system via
overrideLispAttrs, the responsibility (and control over) all of the above is
passed on to the user. Then they can just pick out any subsystems they want.

The one remaining issue is with lispWithPackages wrappers. I have the feeling
that to use overridden packages in the wrapper, one should first overrideScope'
on the package set, then pick out what one wants from the attrset. I think that
calling overrideLispAttrs directly in the list returned by the lispWithPackages
function will not propagate through the scope, and would create multiple
same-named system definition files in the source registry.

To support that we could use the approach from pythonPackages, where there is a
python derivation that has a packageOverrides argument, which updates the scope
of its packages. We could have the same for lisp derivations for each of the
supported implementations (currently ABCL, CCL, Clasp, CLISP, ECL and SBCL).