Don't remove the skeleton.asd in cl-project
cl-project is used for creating new common lisp projects. As such, it
normally creates a .asd file which is generated from the template
skeleton/skeleton.asd. If this file is missing, cl-project is silently
skipping the creation of an .asd file for the new project. Thus there
needs to be an escape hatch from removing all .asd files during the
installPhase without building that .asd file.